Fix codeclimate. Part 5
